pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multiagent-ai-platform"
version = "1.0.0"
description = "Multi-Agent AI Platform with FastAPI, LangGraph, Bedrock, Pinecone and Google Embeddings"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # ── Web Framework ──────────────────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",          # file uploads

    # ── Config & Validation ────────────────────────────────────────────────────
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # ── Database ───────────────────────────────────────────────────────────────
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",                  # async PostgreSQL driver
    "alembic>=1.13.0",                  # migrations

    # ── LangChain / LangGraph ──────────────────────────────────────────────────
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",

    # ── Amazon Bedrock ─────────────────────────────────────────────────────────
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "langchain-aws>=0.2.0",

    # ── Google Vertex AI / Gemini Embeddings ───────────────────────────────────
    "google-cloud-aiplatform>=1.60.0",
    "google-auth>=2.30.0",

    # ── Pinecone ───────────────────────────────────────────────────────────────
    "pinecone-client>=4.0.0",

    # ── AWS S3 (included via boto3) ────────────────────────────────────────────

    # ── Document Parsers ───────────────────────────────────────────────────────
    "pypdf>=4.2.0",                     # PDF parsing
    "python-docx>=1.1.0",               # DOCX parsing
    "markdown>=3.6.0",                  # Markdown
    "pytesseract>=0.3.10",              # OCR (optional)
    "Pillow>=10.3.0",                   # image support for OCR

    # ── Text Splitting ─────────────────────────────────────────────────────────
    "langchain-text-splitters>=0.3.0",

    # ── Utilities ─────────────────────────────────────────────────────────────
    "httpx>=0.27.0",
    "tenacity>=8.3.0",                  # retries
    "structlog>=24.1.0",                # structured logging
    "ulid-py>=1.1.0",                   # sortable unique IDs for chunks
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",                    # async test client
    "factory-boy>=3.3.0",               # test factories
    "faker>=25.0.0",
    "ruff>=0.4.0",                      # linter + formatter
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ── Ruff (linter) ──────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.per-file-ignores]
"tests/*" = ["ANN"]

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
